feat/#27-미리보기 화면 추가
