+++ Caladium/Caladium/ViewModels/CameraViewModel.swift
@@ -6,6 +6,7 @@
 //
 
 import SwiftUI
+import Combine
 
 final class CameraViewModel: ObservableObject {
     
@@ -15,9 +16,27 @@ final class CameraViewModel: ObservableObject {
     @Published var isOverlayOn = false
     @Published var overlayImage: UIImage?
     
+    private var currentContext: CameraContext?
+    private var cancellables = Set<AnyCancellable>()
+    
     init(coordinator: AppCoordinator) {
         self.coordinator = coordinator
+        
+        // CameraService의 capturedImage 변경 감지
+        cameraService.$capturedImage
+            .compactMap { $0 } // nil이 아닌 값만 통과
+            .sink { [weak self] image in
+                print("📸 ViewModel: 새로운 이미지 수신됨")
+                self?.handleCapturedImage(image)
+            }
+            .store(in: &cancellables)
     }
+    
+    func setContext(_ context: CameraContext) {
+        print("🔥 setContext 호출됨: \(context)")
+        self.currentContext = context
+    }
+    
     
     func cancel() {
         print("취소")
@@ -27,10 +46,17 @@ final class CameraViewModel: ObservableObject {
     func capturePhoto() {
         print("촬영")
         cameraService.capturePhoto()
+    }
+    
+    func handleCapturedImage(_ image: UIImage) {
+        print("이미지 처리 시작")
+        self.overlayImage = image
         
-        // 사진 촬영 -> cameraService.capturedImage에 값이 할당 -> overlayImage에 저장
-        if let image = cameraService.capturedImage {
-            self.overlayImage = image
+        if let context = currentContext {
+            print("네비게이션 실행")
+            coordinator.navigate(to: .photoConfirm(image, context))
+        } else {
+            print("context가 nil")
         }
     }
